Refactor(service): Outsource interface definition in a seperate file for better readability, update all importing files aligning to the change
