+++ frontend/src/services/shoppingItem.service.ts
@@ -5,15 +5,6 @@ const API = axios.create({
   withCredentials: true,
 });
 
-// Define the type for a shopping item
-// This type should match the structure of the shopping items in the backend
-export interface ShoppingItemType {
-  _id: string;
-  name: string;
-  bought: boolean;
-  createdAt: Date;
-}
-
 // Get all shopping items
 export const getAllShoppingItems = async () => {
   try {
